More efficient random weighted
Instead of always building the weighted collection, we can shortcut things by just rolling percentile directly to represent the weight. As long as there is a common, this will work. Otherwise, we will build the weighted collection.

This will help speed up randomization when working with very large weighted collections, such as appearances in CreatureGen